# Conflicts:
#	src/main/java/com/ebicep/warlords/classes/abilties/Breath.java
#	src/main/java/com/ebicep/warlords/classes/abilties/Chain.java
#	src/main/java/com/ebicep/warlords/classes/abilties/Fireball.java
#	src/main/java/com/ebicep/warlords/classes/abilties/FrostBolt.java
#	src/main/java/com/ebicep/warlords/classes/abilties/HealingRain.java
#	src/main/java/com/ebicep/warlords/classes/abilties/Intervene.java
#	src/main/java/com/ebicep/warlords/classes/abilties/OrbsOfLife.java
#	src/main/java/com/ebicep/warlords/classes/abilties/Strike.java
#	src/main/java/com/ebicep/warlords/classes/abilties/WaterBolt.java
#	src/main/java/com/ebicep/warlords/classes/abilties/Windfury.java

+++ src/main/java/com/ebicep/warlords/classes/abilties/TimeWarp.java
@@ -112,9 +112,9 @@ public class TimeWarp extends AbstractAbility {
         private WarlordsPlayer warlordsPlayer;
         private Location location;
         private Vector facing;
-        private int time;
+        private float time;
 
-        public TimeWarpPlayer(WarlordsPlayer warlordsPlayer, Location location, Vector facing, int time) {
+        public TimeWarpPlayer(WarlordsPlayer warlordsPlayer, Location location, Vector facing, float time) {
             this.warlordsPlayer = warlordsPlayer;
             this.location = location;
             this.facing = facing;
@@ -145,11 +145,11 @@ public class TimeWarp extends AbstractAbility {
             this.facing = facing;
         }
 
-        public int getTime() {
+        public float getTime() {
             return time;
         }
 
-        public void setTime(int time) {
+        public void setTime(float time) {
             this.time = time;
         }
     }
